Refactor code for better readability and maintainability
Refactor code to improve modularity and maintainability.

* **Add `utils.py`**: Create a utility module for environment variable checks and logging setup.
* **Add `azure_utils.py`**: Create a module for Azure-related functions, including deleting existing PDFs, creating search resources, resetting and running the indexer, and getting indexer status.
* **Add `config.py`**: Create a configuration module for FastAPI app and middleware setup.
* **Modify `main.py`**: Remove environment variable checks, logging setup, Azure-related functions, and FastAPI app and middleware setup. Import and use functions from `utils.py`, `azure_utils.py`, and `config.py`.
* **Update `requirements.txt`**: Review and update dependencies for compatibility and security.
