+++ requirements.txt
@@ -9,5 +9,3 @@ python-dotenv
 pydantic==2.4.2
 streamlit>=1.22.0
 requests>=2.28.2
-# Add other dependencies your app needs
-
